requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
moviepy==1.0.3
pydub==0.25.1
werkzeug==3.0.1
ffmpeg-python==0.2.0
torch>=2.0.0
numpy>=1.24.0